+++ api/src/main/java/me/m56738/easyarmorstands/api/editor/node/EmptyNodeImpl.java
@@ -1,0 +1,5 @@
+package me.m56738.easyarmorstands.api.editor.node;
+
+class EmptyNodeImpl implements Node {
+    public static final EmptyNodeImpl INSTANCE = new EmptyNodeImpl();
+}