+++ src/main/java/me/m56738/easyarmorstands/editor/display/layer/DisplayRootLayer.java
@@ -5,17 +5,15 @@ import me.m56738.easyarmorstands.api.editor.context.ClickContext;
 import me.m56738.easyarmorstands.api.editor.context.UpdateContext;
 import me.m56738.easyarmorstands.api.editor.layer.ElementLayer;
 import me.m56738.easyarmorstands.api.editor.layer.ResettableLayer;
-import me.m56738.easyarmorstands.editor.util.ToolMenuManager;
 import me.m56738.easyarmorstands.api.element.Element;
 import me.m56738.easyarmorstands.api.property.Property;
-import me.m56738.easyarmorstands.api.property.type.EntityPropertyTypes;
-import me.m56738.easyarmorstands.command.sender.EasPlayer;
 import me.m56738.easyarmorstands.api.property.type.BlockDisplayPropertyTypes;
 import me.m56738.easyarmorstands.api.property.type.DisplayPropertyTypes;
-import me.m56738.easyarmorstands.element.DisplayElement;
+import me.m56738.easyarmorstands.api.property.type.EntityPropertyTypes;
+import me.m56738.easyarmorstands.command.sender.EasPlayer;
 import me.m56738.easyarmorstands.editor.input.OpenElementMenuInput;
 import me.m56738.easyarmorstands.editor.input.ReturnInput;
-import me.m56738.easyarmorstands.editor.layer.ToolMenuModeSwitcher;
+import me.m56738.easyarmorstands.element.DisplayElement;
 import me.m56738.easyarmorstands.message.Message;
 import me.m56738.easyarmorstands.permission.Permissions;
 import org.bukkit.Location;
@@ -31,8 +29,6 @@ public class DisplayRootLayer extends DisplayLayer implements ElementLayer, Rese
     private final DisplayElement<?> element;
     private final Property<Location> locationProperty;
     private final Property<BlockData> blockDataProperty;
-    private final ToolMenuManager toolManager;
-    private final ToolMenuModeSwitcher toolSwitcher;
     private final boolean allowMenu;
 
     public DisplayRootLayer(Session session, DisplayElement<?> element) {
@@ -41,9 +37,8 @@ public class DisplayRootLayer extends DisplayLayer implements ElementLayer, Rese
         this.element = element;
         this.locationProperty = properties().get(EntityPropertyTypes.LOCATION);
         this.blockDataProperty = properties().getOrNull(BlockDisplayPropertyTypes.BLOCK);
-        this.toolManager = new ToolMenuManager(session, this, element.getTools(properties()));
-        this.toolSwitcher = new ToolMenuModeSwitcher(this.toolManager);
         this.allowMenu = session.player().hasPermission(Permissions.OPEN);
+        addNode(session.nodeProvider().tools(element.getTools(properties())));
     }
 
     @Override
@@ -69,8 +64,6 @@ public class DisplayRootLayer extends DisplayLayer implements ElementLayer, Rese
 
     @Override
     public void onUpdate(@NotNull UpdateContext context) {
-        context.setActionBar(toolSwitcher.getActionBar());
-        toolSwitcher.onUpdate(context);
         if (allowMenu) {
             context.addInput(new OpenElementMenuInput(session, element));
         }