+++ src/main/java/me/m56738/easyarmorstands/editor/layer/SimpleEntityLayer.java
@@ -2,10 +2,8 @@ package me.m56738.easyarmorstands.editor.layer;
 
 import me.m56738.easyarmorstands.api.editor.Session;
 import me.m56738.easyarmorstands.api.editor.context.UpdateContext;
-import me.m56738.easyarmorstands.api.editor.layer.ElementLayer;
 import me.m56738.easyarmorstands.api.editor.layer.AbstractLayer;
-import me.m56738.easyarmorstands.editor.util.ToolMenuManager;
-import me.m56738.easyarmorstands.editor.util.ToolMenuMode;
+import me.m56738.easyarmorstands.api.editor.layer.ElementLayer;
 import me.m56738.easyarmorstands.api.element.EditableElement;
 import me.m56738.easyarmorstands.api.element.Element;
 import me.m56738.easyarmorstands.api.element.MenuElement;
@@ -17,24 +15,18 @@ import org.jetbrains.annotations.NotNull;
 public class SimpleEntityLayer extends AbstractLayer implements ElementLayer {
     private final Session session;
     private final Element element;
-    private final ToolMenuManager toolManager;
-    private final ToolMenuModeSwitcher toolSwitcher;
     private final boolean allowMenu;
 
     public SimpleEntityLayer(Session session, EditableElement element) {
         super(session);
         this.session = session;
         this.element = element;
-        this.toolManager = new ToolMenuManager(session, this, element.getTools(session.properties(element)));
-        this.toolManager.setMode(ToolMenuMode.GLOBAL);
-        this.toolSwitcher = new ToolMenuModeSwitcher(this.toolManager);
         this.allowMenu = session.player().hasPermission(Permissions.OPEN);
+        addNode(session.nodeProvider().tools(element.getTools(session.properties(element))));
     }
 
     @Override
     public void onUpdate(@NotNull UpdateContext context) {
-        context.setActionBar(toolSwitcher.getActionBar());
-        toolSwitcher.onUpdate(context);
         if (allowMenu && element instanceof MenuElement) {
             context.addInput(new OpenElementMenuInput(session, (MenuElement) element));
         }