+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Improvements
 
 * Updated to 26.2
+* Simplified tool mode display
+
+# API changes
+
+* Added tools to NodeProvider